fix: fix Group Missing Exceptions (#129)
In the case where groups are not returned by an OIDC provider, this code
previously defaulted the group result to an empty list. This fails to
raise the appropriate "Failed to get user groups" error and instead
continues on before failing and returning a more confusing error.

This fixes that and raises the appropriate error if groups are not
returned.